+++ spec/spec_helper.rb
@@ -8,13 +8,22 @@ SimpleCov.start do
   add_filter 'lib/tiktok/open/sdk/version'
 end
 
+require 'webmock/rspec'
+require 'omniauth'
+require 'omniauth-oauth2'
 require 'tiktok/open/sdk'
 
-RSpec.configure do |config|
-  # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = '.rspec_status'
+Tiktok::Open::Sdk.configure do |config|
+  config.client_key             = 'test_client_key'
+  config.client_secret          = 'test_client_secret'
+  config.user_auth.scopes       = %w[user.info.basic user.info.profile]
+  config.user_auth.redirect_uri = 'https://example.com/auth/tiktok_open_sdk/callback'
+end
 
-  # Disable RSpec exposing methods globally on `Module` and `main`
+require 'tiktok/open/omniauth/strategies/tiktok_open_sdk'
+
+RSpec.configure do |config|
+  config.example_status_persistence_file_path = '.rspec_status'
   config.disable_monkey_patching!
 
   config.expect_with :rspec do |c|